+++ stories/User Interaction/Date and Time Input/DateTimePicker.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/nextjs'
-import { DateTimePickerUncontrolled } from '../../../src/components/user-interaction/date/DateAndTimePicker'
+import { DateTimePickerUncontrolled } from '../../../src/components/user-interaction/date/DateTimePicker'
 import { addDuration, subtractDuration } from '../../../src/utils/date'
 import { action } from 'storybook/actions'
 
@@ -13,14 +13,15 @@ type Story = StoryObj<typeof meta>;
 
 export const dateTimePicker: Story = {
   args: {
-    mode: 'dateTime',
     value: new Date(),
+    mode: 'dateTime',
     start: subtractDuration(new Date(), { years: 50 }),
     end: addDuration(new Date(), { years: 50 }),
     is24HourFormat: true,
     minuteIncrement: '5min',
     weekStart: 'monday',
     markToday: true,
-    onValueChange: action('onValueChange')
+    onValueChange: action('onValueChange'),
+    onEditComplete: action('onEditComplete')
   },
 }