+++ stories/User Interaction/Date and Time Input/DayPicker.stories.tsx
@@ -12,10 +12,11 @@ type Story = StoryObj<typeof meta>;
 export const dayPicker: Story = {
   args: {
     displayedMonth: new Date(),
-    selected: new Date(),
+    value: new Date(),
     markToday: true,
     weekStart: 'monday',
     className: 'h-max-71',
     onValueChange: action('onValueChange'),
+    onEditComplete: action('onEditComplete'),
   },
 }