+++ stories/User Interaction/Date and Time Input/TimePicker.stories.tsx
@@ -11,10 +11,17 @@ type Story = StoryObj<typeof meta>;
 
 export const timePicker: Story = {
   args: {
-    time: new Date(),
+    value: new Date(),
     is24HourFormat: true,
     minuteIncrement: '5min',
-    maxHeight: 300,
-    onValueChange: action('onValueChange')
+    onValueChange: action('onValueChange'),
+    onEditComplete: action('onEditComplete')
   },
+  decorators: (Story) => {
+    return (
+      <div className="flex flex-col overflow-hidden h-64">
+        <Story />
+      </div>
+    )
+  }
 }